refactor: streamline authentication context and loading states in layout components
